chore: remove tsx
