+++ cspell-words.txt
@@ -4,8 +4,9 @@ fullstacksjs
 gitmessage
 grammyjs
 libsql
-listrepos
 listcontributors
+listrepos
+nodenext
 removerepo
-smee
 Safaeirad
+smee
