+++ drizzle.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from "drizzle-kit";
 
-import { config } from "./src/config";
+import { config } from "./src/config/index.ts";
 
 export default defineConfig({
   dialect: "sqlite",
