+++ package.json
@@ -3,16 +3,36 @@
   "type": "module",
   "private": true,
   "packageManager": "pnpm@10.22.0",
+  "imports": {
+    "#config": [
+      "./src/config/index.ts"
+    ],
+    "#db": [
+      "./src/db/index.ts"
+    ],
+    "#bot": [
+      "./src/bot/bot.ts"
+    ],
+    "#github": [
+      "./src/lib/github/index.ts"
+    ],
+    "#webhooks": [
+      "./src/lib/github/webhooks/index.ts"
+    ],
+    "#telegram": [
+      "./src/lib/telegram.ts"
+    ]
+  },
   "scripts": {
-    "dev": "node --watch --import=tsx --env-file-if-exists=.env src/index.ts",
-    "start": "node --import=tsx --env-file-if-exists=.env src/index.ts",
+    "dev": "node --watch --env-file-if-exists=.env src/index.ts",
+    "start": "node --env-file-if-exists=.env src/index.ts",
     "lint": "eslint . --fix",
     "typecheck": "tsc",
     "smee": "smee",
     "db:studio": "drizzle-kit studio",
     "db:migrate": "drizzle-kit migrate",
     "db:generate": "drizzle-kit generate",
-    "gh:discover": "node --import=tsx --env-file-if-exists=.env scripts/gh-discover.ts",
+    "gh:discover": "node --env-file-if-exists=.env scripts/gh-discover.ts",
     "prepare": "simple-git-hooks && git config --local commit.template '.gitmessage'"
   },
   "dependencies": {
@@ -46,7 +66,6 @@
     "prettier": "3.6.2",
     "simple-git-hooks": "2.13.1",
     "smee-client": "5.0.0",
-    "tsx": "4.20.6",
     "typescript": "5.9.3"
   }
 }
