+++ src/bot/commands/index.ts
@@ -1,16 +1,16 @@
 import { CommandGroup } from "@grammyjs/commands";
 
-import type { BotContext } from "../bot";
+import type { BotContext } from "../bot.ts";
 
-import { cmdAddRepo } from "./addrepo";
-import { cmdDiscover } from "./discover";
-import { cmdHelp } from "./help";
-import { cmdLink } from "./link";
-import { cmdListContributors } from "./listcontributors";
-import { cmdListRepos } from "./listrepos";
-import { cmdRemoveRepo } from "./removerepo";
-import { cmdUnLink } from "./unlink";
-import { cmdWhoami } from "./whoami";
+import { cmdAddRepo } from "./addrepo.ts";
+import { cmdDiscover } from "./discover.ts";
+import { cmdHelp } from "./help.ts";
+import { cmdLink } from "./link.ts";
+import { cmdListContributors } from "./listcontributors.ts";
+import { cmdListRepos } from "./listrepos.ts";
+import { cmdRemoveRepo } from "./removerepo.ts";
+import { cmdUnLink } from "./unlink.ts";
+import { cmdWhoami } from "./whoami.ts";
 
 export const commands = new CommandGroup<BotContext>().add([
   cmdAddRepo,
