+++ src/bot/commands/listrepos.ts
@@ -1,11 +1,10 @@
 import { Command } from "@grammyjs/commands";
+import { config } from "#config";
+import { db } from "#db";
 
-import { config } from "@/config";
-import { db } from "@/db";
+import type { BotContext } from "../bot.ts";
 
-import type { BotContext } from "../bot";
-
-import { escapeMarkdown } from "../../lib/escape-markdown";
+import { escapeMarkdown } from "../../lib/escape-markdown.ts";
 
 export async function listreposHandler(ctx: BotContext) {
   if (!ctx.message) return;
