+++ src/bot/commands/removerepo.ts
@@ -1,13 +1,12 @@
 import { Command } from "@grammyjs/commands";
+import { config } from "#config";
+import { db, schema } from "#db";
+import { isGitHubUrl } from "#github";
 import { eq } from "drizzle-orm";
 
-import { config } from "@/config";
-import { db, schema } from "@/db";
-import { isGitHubUrl } from "@/github";
+import type { BotContext } from "../bot.ts";
 
-import type { BotContext } from "../bot";
-
-import { escapeMarkdown } from "../../lib/escape-markdown";
+import { escapeMarkdown } from "../../lib/escape-markdown.ts";
 
 export async function removerepoHandler(ctx: BotContext) {
   if (!ctx.message) return;
