+++ src/db/schema/index.ts
@@ -1,2 +1,2 @@
-export * from "./relations";
-export * from "./tables";
+export * from "./relations.ts";
+export * from "./tables.ts";
