+++ src/lib/github/index.ts
@@ -1,0 +1,1 @@
+export * from "./github.ts";
